Classifier/requirements.txt
Keras==2.4.3
Keras-Preprocessing==1.1.2
scikit-image==0.17.2
scikit-learn==0.23.2
numpy==1.18.5
opencv-python==4.4.0.46
Pillow==8.0.1
